[Feature] 반응형 기초 세팅 #11
