+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { RecoilRoot } from 'recoil';
+import { RecoilRoot, useSetRecoilState } from 'recoil';
 import { useRouter } from 'next/router';
 
 import LayoutContainer from '../components/layouts';
@@ -11,13 +11,7 @@ import 'slick-carousel/slick/slick-theme.css';
 function MyApp({ Component, pageProps }: any) {
 	const router = useRouter();
 
-	const checkViewPortWidth = () => {
-		if (window.innerWidth <= 640) {
-			document.documentElement.style.fontSize = '5vw';
-		} else {
-			document.documentElement.style.fontSize = '';
-		}
-	};
+	// error debug
 	useEffect(() => {
 		const originalError = console.error;
 		// ...args : 함수에 전달되는 모든 인자들
@@ -33,10 +27,7 @@ function MyApp({ Component, pageProps }: any) {
 		};
 	}, []);
 
-	useEffect(() => {
-		window.addEventListener('resize', checkViewPortWidth);
-	}, []);
-
+	// router replace
 	useEffect(() => {
 		if (router.pathname === '/' || router.pathname === '/_error') {
 			router.replace('/main');
